Fix rat check and rename some files
