+++ stream-client/python/streamclient/TESTS.md
@@ -13,13 +13,13 @@ self.config_file = 'config_name.json'
 ```
 
 In command shell type this:
-```> python runtests_reactor.py```
-```> python runtests_reactor_ssl.py```
+```> python stream_integration_test.py```
+```> python stream_integration_test_ssl.py```
 
 ## To configure tests change these files:
 ```config.json```                  - for unit tests.
-```config_reactor.json```          - for integration with Singlenode reactor
-```config_reactor.json```          - for integration with SSL and Auth enabled reactor
+```cdap_config.json```             - for integration with Singlenode reactor
+```cdap_ssl_config.json```         - for integration with SSL and Auth enabled reactor
 
 ## More server instances to tests.
 New tests should be derived from *BasicReactor* and *unittest.TestCase* classes.
@@ -29,9 +29,25 @@ and call *base_set_up* of *BasicReactor* to initialize tests.
 Example:
 
 ```
-class TestStreamClient(unittest.TestCase, BasicReactor):
+class TestStreamClient(unittest.TestCase, StreamTestBase):
 
     def setUp(self):
         self.config_file = 'config_reactor.json'
         self.base_set_up()
 ```
+
+#
+# Copyright © 2014 Cask Data, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"); you may not
+# use this file except in compliance with the License. You may obtain a copy of
+# the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations under
+# the License.
+#
