+++ stream-client/python/streamclient/v2/test/stream_test_base.py
@@ -1,3 +1,19 @@
+# -*- coding: utf-8 -*-
+
+#  Copyright © 2014 Cask Data, Inc.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License"); you may not
+#  use this file except in compliance with the License. You may obtain a copy of
+#  the License at
+#
+#  http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+#  WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+#  License for the specific language governing permissions and limitations under
+#  the License.
+
 import requests
 
 import os
@@ -19,7 +35,7 @@ from caskauthclient.Config import Config as AuthConfig
 # In children 'config_file' property has to be set and
 # 'base_set_up' method called.
 
-class BasicReactor(object):
+class StreamTestBase(object):
 
 
     validStream = u'validStream'
@@ -154,7 +170,7 @@ class BasicReactor(object):
             self.exit_code = response.status_code
 
         def check_exit_code(response):
-                self.assertEqual(self.exit_code, 200)
+            self.assertEqual(self.exit_code, 200)
 
         q = sw.send(self.validFile)
         q.on_response(on_response)
@@ -167,7 +183,7 @@ class BasicReactor(object):
             self.exit_code = response.status_code
 
         def check_exit_code(response):
-                self.assertEqual(self.exit_code, 200)
+            self.assertEqual(self.exit_code, 200)
 
         q = sw.write(self.messageToWrite)
         q.on_response(on_response)
