--- stream-client/python/streamclient/v3/test/runtests_reactor_ssl.py
@@ -1,17 +1,0 @@
-#! /usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-import unittest
-import requests
-
-from basicreactor import BasicReactor
-
-
-class TestStreamClient(unittest.TestCase, BasicReactor):
-
-    def setUp(self):
-        self.config_file = 'config_reactor_ssl.json'
-        self.base_set_up()
-
-if '__main__' == __name__:
-    unittest.main(warnings=False)
