+++ stream-client/python/streamclient/v3/test/stream_integration_test.py
@@ -1,3 +1,4 @@
+#! /usr/bin/env python3
 # -*- coding: utf-8 -*-
 
 #  Copyright © 2014 Cask Data, Inc.
@@ -13,3 +14,18 @@
 #  WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 #  License for the specific language governing permissions and limitations under
 #  the License.
+
+import unittest
+import requests
+
+from basicreactor import BasicReactor
+
+
+class TestStreamClient(unittest.TestCase, StreamTestBase):
+
+    def setUp(self):
+        self.config_file = 'cdap_config.json'
+        self.base_set_up()
+
+if '__main__' == __name__:
+    unittest.main(warnings=False)
